Update resume.qmd
You have a well-built resume and I couldn't find ways to improve it. I did reword your experience section, it is not necessarily making it better,  but just different options to write down some of the sentences so you can pick if you want to change them, or just one or none.